+++ sdk/include/opentelemetry/sdk/configuration/configuration_parser.h
@@ -23,6 +23,7 @@
 #include "opentelemetry/sdk/configuration/console_span_exporter_configuration.h"
 #include "opentelemetry/sdk/configuration/default_aggregation_configuration.h"
 #include "opentelemetry/sdk/configuration/default_histogram_aggregation.h"
+#include "opentelemetry/sdk/configuration/distribution_configuration.h"
 #include "opentelemetry/sdk/configuration/document.h"
 #include "opentelemetry/sdk/configuration/document_node.h"
 #include "opentelemetry/sdk/configuration/double_array_attribute_value_configuration.h"
@@ -363,6 +364,9 @@ public:
       const std::unique_ptr<DocumentNode> &node) const;
 
   std::unique_ptr<ResourceConfiguration> ParseResourceConfiguration(
+      const std::unique_ptr<DocumentNode> &node) const;
+
+  std::unique_ptr<DistributionConfiguration> ParseDistributionConfiguration(
       const std::unique_ptr<DocumentNode> &node) const;
 
   std::unique_ptr<Configuration> Parse(std::unique_ptr<Document> doc);