+++ sdk/src/configuration/configuration_parser.cc
@@ -31,6 +31,8 @@
 #include "opentelemetry/sdk/configuration/console_span_exporter_configuration.h"
 #include "opentelemetry/sdk/configuration/default_aggregation_configuration.h"
 #include "opentelemetry/sdk/configuration/default_histogram_aggregation.h"
+#include "opentelemetry/sdk/configuration/distribution_configuration.h"
+#include "opentelemetry/sdk/configuration/distribution_entry_configuration.h"
 #include "opentelemetry/sdk/configuration/document.h"
 #include "opentelemetry/sdk/configuration/document_node.h"
 #include "opentelemetry/sdk/configuration/double_array_attribute_value_configuration.h"
@@ -2144,6 +2146,33 @@ std::unique_ptr<ResourceConfiguration> ConfigurationParser::ParseResourceConfigu
   return model;
 }
 
+std::unique_ptr<DistributionConfiguration> ConfigurationParser::ParseDistributionConfiguration(
+    const std::unique_ptr<DocumentNode> &node) const
+{
+  auto model = std::make_unique<DistributionConfiguration>();
+
+  for (auto it = node->begin(); it != node->end(); ++it)
+  {
+    std::unique_ptr<DocumentNode> child(*it);
+    std::string name = child->Key();
+
+    auto entry  = std::make_unique<DistributionEntryConfiguration>();
+    entry->name = std::move(name);
+    entry->node = std::move(child);
+
+    model->entries.push_back(std::move(entry));
+  }
+
+  size_t count = model->entries.size();
+  if (count == 0)
+  {
+    std::string message("Illegal distribution, 0 entries");
+    throw InvalidSchemaException(node->Location(), message);
+  }
+
+  return model;
+}
+
 std::unique_ptr<Configuration> ConfigurationParser::Parse(std::unique_ptr<Document> doc)
 {
   std::unique_ptr<DocumentNode> node = doc->GetRootNode();
@@ -2229,7 +2258,11 @@ std::unique_ptr<Configuration> ConfigurationParser::Parse(std::unique_ptr<Docume
 
   // FIXME: instrumentation/development
 
-  // FIXME: distribution
+  child = node->GetChildNode("distribution");
+  if (child)
+  {
+    model->distribution = ParseDistributionConfiguration(child);
+  }
 
   return model;
 }